checked parallel version

+++ count_number_of_all_words/includes/parallel_program.h
@@ -13,6 +13,6 @@ void count_words(std::vector<std::string> &data, const int start_position, const
                  std::map<std::string, int> &map_of_words);
 
 void parallel_count(const std::string &input_filename, const std::string &output_filename_a,
-                    const std::string &output_filename_n, const int num_threads);
+                    const std::string &output_filename_n, const uint8_t num_threads);
 
 #endif //ARCHITECTURE_OF_COMPUTER_SYSTEMS_PARALLEL_PROGRAM_H
